convert api project to ts

See merge request kidasov/competition!10

+++ api/src/routes/events/events.ts
@@ -1,9 +1,9 @@
-import Router from 'koa-router';
-import { Event, Attendee, User } from '../../db/models';
+import * as t from 'io-ts';
+import * as Router from 'koa-router';
+import * as Sequelize from 'sequelize';
+import { Attendee, Event, User } from '../../db/models';
+import { asEventId, PublishState } from '../../db/models/event';
 import attendees from './attendees';
-import { DRAFT, PUBLISHED } from '../../db/models/event';
-import { sequelize } from '../../db/db';
-import Sequelize from 'sequelize';
 
 const Op = Sequelize.Op;
 
@@ -15,22 +15,31 @@ router.use(
 );
 
 router.get('/', async ctx => {
-  const { userId } = ctx.session || {};
+  const sessionUserId = await ctx.sessionUserId();
 
   ctx.status = 200;
   ctx.body = await Event.findAll({
-    where: { [Op.or]: [{ state: PUBLISHED }, { ownerUserId: userId }] },
+    where: {
+      [Op.or]: [
+        { state: PublishState.Published },
+        { ownerUserId: sessionUserId },
+      ],
+    },
   });
 });
 
+const CreateEventRequest = t.type({
+  name: t.string,
+});
+
 router.post('/', async ctx => {
-  const { userId } = ctx.requireSession();
-  const { name } = ctx.request.body;
+  const { userId: sessionUserId } = await ctx.requireSession();
+  const { name } = ctx.decode(CreateEventRequest);
 
   const event = await Event.create({
     name,
-    ownerUserId: userId,
-    state: DRAFT,
+    ownerUserId: sessionUserId,
+    state: PublishState.Draft,
   });
 
   ctx.status = 201;
@@ -38,18 +47,23 @@ router.post('/', async ctx => {
 });
 
 router.delete('/:id', async ctx => {
-  const { id } = ctx.params;
+  const id = asEventId(ctx.paramNumber('id'));
+
   const event = await Event.findOne({
     where: { id },
   });
 
-  await Event.destroy({ where: { id: event.id } });
+  if (event) {
+    await Event.destroy({ where: { id: event.id } });
+  }
+
   ctx.status = 200;
   ctx.body = {};
 });
 
 router.get('/:id', async ctx => {
-  const { id } = ctx.params;
+  const id = asEventId(ctx.paramNumber('id'));
+
   const event = await Event.findOne({
     where: { id },
   });
@@ -73,10 +87,18 @@ router.get('/:id', async ctx => {
   ctx.status = 200;
 });
 
+const PatchEventRequest = t.type({
+  name: t.string,
+  state: t.union([
+    t.literal(PublishState.Draft),
+    t.literal(PublishState.Published),
+  ]),
+});
+
 router.patch('/:id', async ctx => {
-  const { userId } = ctx.requireSession();
-  const { id } = ctx.params;
-  const { name, state } = ctx.request.body;
+  const { userId: sessionUserId } = await ctx.requireSession();
+  const id = asEventId(ctx.paramNumber('id'));
+  const { name, state } = ctx.decode(PatchEventRequest);
 
   const event = await Event.findOne({ where: { id } });
 
@@ -84,7 +106,7 @@ router.patch('/:id', async ctx => {
     return ctx.throw(404);
   }
 
-  if (userId !== event.ownerUserId) {
+  if (sessionUserId !== event.ownerUserId) {
     return ctx.throw(403);
   }
